+++ src/Workspace.cpp
@@ -905,25 +905,6 @@ lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
     return lsp::WorkspaceEdit{{{params.textDocument.uri.toString(), localChanges}}};
 }
 
-// std::optional<std::vector<lsp::DocumentSymbol>> WorkspaceFolder::documentSymbol(const lsp::DocumentSymbolParams& params)
-// {
-//     auto moduleName = getModuleName(params.textDocument.uri);
-
-//     // Run the type checker to ensure we are up to date
-//     if (frontend.isDirty(moduleName))
-//         frontend.check(moduleName);
-
-//     auto module = frontend.moduleResolver.getModule(moduleName);
-//     if (!module)
-//         return std::nullopt;
-
-//     std::vector<lsp::DocumentSymbol> result;
-
-//     // TODO
-
-//     return result;
-// }
-
 bool WorkspaceFolder::updateSourceMap()
 {
     // Read in the sourcemap